+++ mux/src/ssh.rs
@@ -996,6 +996,11 @@ impl portable_pty::MasterPty for WrappedSshPty {
     fn as_raw_fd(&self) -> Option<std::os::fd::RawFd> {
         None
     }
+
+    #[cfg(unix)]
+    fn tty_name(&self) -> Option<std::path::PathBuf> {
+        None
+    }
 }
 
 impl std::io::Write for PtyWriter {