+++ pty/src/serial.rs
@@ -17,6 +17,7 @@ use serial::{
 use std::cell::RefCell;
 use std::ffi::{OsStr, OsString};
 use std::io::{Read, Result as IoResult, Write};
+use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
@@ -243,6 +244,11 @@ impl MasterPty for Master {
 
     #[cfg(unix)]
     fn as_raw_fd(&self) -> Option<crate::unix::RawFd> {
+        None
+    }
+
+    #[cfg(unix)]
+    fn tty_name(&self) -> Option<PathBuf> {
         None
     }
 }