+++ pty/src/unix.rs
@@ -5,9 +5,12 @@ use anyhow::{bail, Error};
 use filedescriptor::FileDescriptor;
 use libc::{self, winsize};
 use std::cell::RefCell;
+use std::ffi::OsStr;
 use std::io::{Read, Write};
+use std::os::unix::ffi::OsStrExt;
 use std::os::unix::io::{AsRawFd, FromRawFd};
 use std::os::unix::process::CommandExt;
+use std::path::PathBuf;
 use std::{io, mem, ptr};
 
 pub use std::os::unix::io::RawFd;
@@ -42,9 +45,12 @@ fn openpty(size: PtySize) -> anyhow::Result<(UnixMasterPty, UnixSlavePty)> {
         bail!("failed to openpty: {:?}", io::Error::last_os_error());
     }
 
+    let tty_name = tty_name(slave);
+
     let master = UnixMasterPty {
         fd: PtyFd(unsafe { FileDescriptor::from_raw_fd(master) }),
         took_writer: RefCell::new(false),
+        tty_name,
     };
     let slave = UnixSlavePty {
         fd: PtyFd(unsafe { FileDescriptor::from_raw_fd(slave) }),
@@ -95,6 +101,33 @@ impl Read for PtyFd {
             }
             x => x,
         }
+    }
+}
+
+fn tty_name(fd: RawFd) -> Option<PathBuf> {
+    let mut buf = vec![0 as std::ffi::c_char; 128];
+
+    loop {
+        let res = unsafe { libc::ttyname_r(fd, buf.as_mut_ptr(), buf.len()) };
+
+        if res == libc::ERANGE {
+            if buf.len() > 64 * 1024 {
+                // on macOS, if the buf is "too big", ttyname_r can
+                // return ERANGE, even though that is supposed to
+                // indicate buf is "too small".
+                return None;
+            }
+            buf.resize(buf.len() * 2, 0 as std::ffi::c_char);
+            continue;
+        }
+
+        return if res == 0 {
+            let cstr = unsafe { std::ffi::CStr::from_ptr(buf.as_ptr()) };
+            let osstr = OsStr::from_bytes(cstr.to_bytes());
+            Some(PathBuf::from(osstr))
+        } else {
+            None
+        };
     }
 }
 
@@ -269,6 +302,7 @@ impl PtyFd {
 struct UnixMasterPty {
     fd: PtyFd,
     took_writer: RefCell<bool>,
+    tty_name: Option<PathBuf>,
 }
 
 /// Represents the slave end of a pty.
@@ -330,6 +364,10 @@ impl MasterPty for UnixMasterPty {
 
     fn as_raw_fd(&self) -> Option<RawFd> {
         Some(self.fd.0.as_raw_fd())
+    }
+
+    fn tty_name(&self) -> Option<PathBuf> {
+        self.tty_name.clone()
     }
 
     fn process_group_leader(&self) -> Option<libc::pid_t> {