+++ wezterm-ssh/src/pty.rs
@@ -82,6 +82,11 @@ impl portable_pty::MasterPty for SshPty {
     fn as_raw_fd(&self) -> Option<std::os::fd::RawFd> {
         None
     }
+
+    #[cfg(unix)]
+    fn tty_name(&self) -> Option<std::path::PathBuf> {
+        None
+    }
 }
 
 #[derive(Debug)]